+++ docs/conf.py
@@ -54,8 +54,9 @@ source_suffix = '.rst'
 master_doc = 'index'
 
 # General information about the project.
+import datetime
 project = u'docx2csv'
-copyright = u'2018, Ivan Begtin'
+copyright = u'%d, Ivan Begtin' % datetime.datetime.now().year
 
 # The version info for the project you're documenting, acts as replacement
 # for |version| and |release|, also used in various other places throughout
@@ -274,4 +275,4 @@ texinfo_documents = [
 #texinfo_no_detailmenu = False
 
 # sphinx.ext.intersphinx confs
-intersphinx_mapping = {'python': ('https://docs.python.org/2', None)}
+intersphinx_mapping = {'python': ('https://docs.python.org/3', None)}