+++ docx2csv/__init__.py
@@ -1,8 +1,10 @@
 # -*- coding: utf-8 -*-
-__version__ = '0.1.2'
+__version__ = '0.1.3'
 __author__ = "Ivan Begtin (ivan@begtin.tech)"
 __license__ = "BSD"
 
 
 from .converter import extract, extract_tables, analyze
 
+
+__all__ = ["extract", "extract_tables", "analyze"]